% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_hierarchy}
\alias{get_hierarchy}
\alias{get_id}
\alias{get_label}
\alias{get_parent_id}
\alias{get_parent_label}
\alias{get_children_id}
\alias{get_children_label}
\title{Querying Sensing Clues' ontologies}
\usage{
get_hierarchy(url = "https://focus.sensingclues.org/", lang = "en")

get_id(label, hierarchy)

get_label(id, hierarchy)

get_parent_id(id, hierarchy)

get_parent_label(label, hierarchy)

get_children_id(id, hierarchy)

get_children_label(label, hierarchy)
}
\arguments{
\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}

\item{label}{Textual concept definiton, for example \verb{Human-wildlife conflict}.}

\item{hierarchy}{Object retrieved by \code{\link[=get_hierarchy]{get_hierarchy()}}.}

\item{id}{Concept identification, for example \verb{https://sensingclues.poolparty.biz/SCCSSOntology/106}.}
}
\value{
Nested list representing multiple ontologies.
}
\description{
Querying Sensing Clues' ontologies
}
\examples{
\dontrun{
lst <- get_hierarchy() # you have to be logged in to https://focus.sensingclues.org/ to get access

# inspect a single concept
cncpt <- lst$concepts[[1]]

# look up the ID for a given concept label
get_id("Human-wildlife conflict", lst)

# look up the label for a given concept ID
get_label("https://sensingclues.poolparty.biz/SCCSSOntology/106", lst)

# get the parent ID/label for a given concept ID/label
get_parent_id("https://sensingclues.poolparty.biz/SCCSSOntology/106", lst)
get_parent_label("Human-wildlife conflict", lst)

# get the children ID/label(s) for a given concept ID/label
get_children_id("https://sensingclues.poolparty.biz/SCCSSOntology/106", lst)
get_children_label("Human-wildlife conflict", lst)
}
}
