% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Retrieve observation data}
\usage{
get_observations(
  cookie,
  group,
  bounds = NULL,
  from = Sys.Date() - 30,
  to = Sys.Date(),
  aoi = "",
  filteredConcepts = NULL,
  updateProgress = NULL,
  allAttributes = FALSE,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with \code{\link[=get_groups]{get_groups()}}.}

\item{bounds}{Bounding box coordinates (latitude and longitude) in list(north, east, south, west) format. For example \code{list(north=15, east=10, south=-25, west=50)}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{aoi}{Area of interest.}

\item{filteredConcepts}{One or multiple concept definitions, for example \verb{https://sensingclues.poolparty.biz/SCCSSOntology/186}. See \url{https://sensingclues.poolparty.biz/GraphViews/} for all available concepts.}

\item{updateProgress}{A function to update a progress bar object, default is NULL.}

\item{allAttributes}{A boolean. Allows you to collect more attributes for the observations.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}
}
\value{
A data frame with all observations collected by the defined group(s), within the given date range.
}
\description{
Retrieve observation data
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_observations(cookie, group = 'focus-project-1234') # demo group
}
}
