% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_counts.R
\name{get_track_counts}
\alias{get_track_counts}
\title{Frequency of track concepts}
\usage{
get_track_counts(
  cookie,
  group,
  from = "1900-01-01",
  to = "2999-12-31",
  hierarchy = NULL,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with \code{\link[=get_groups]{get_groups()}}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{hierarchy}{A Sensing Clues knowledge graph object, to be retrieved with \code{\link[=get_hierarchy]{get_hierarchy()}}.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}
}
\value{
A data frame of track concept definitions and their frequencies collected by the defined group(s), within the given date range. See \url{https://sensingclues.poolparty.biz/GraphViews/} for all available concepts.
}
\description{
Frequency of track concepts
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_track_counts(cookie, group = "focus-project-1234") # demo group
}
}
