% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_agent}
\alias{get_agent}
\title{Retrieve agent details}
\usage{
get_agent(
  cookie,
  agentId,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{agentId}{Identification character string of a single agent. You can find the agent id of your group members in \href{https://focus.sensingclues.org/}{Focus}.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}
}
\value{
A list with three elements:
\itemize{
\item \code{agentId}: Identifier of the agent.
\item \code{agentName}: Name of the agent, such as the username.
\item \code{description}: Free-text description of the agent.
}

Returns \code{NULL} with a warning if you do not have access to the agent or if the
\code{agentId} is incorrect.
}
\description{
Looks up the details of a single agent, for example to put a name to the agent
identifiers found in track data.
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
lst <- get_agent(cookie, agentId = 'YOUR AGENT ID')
}
}
