% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups}
\alias{get_groups}
\title{Overview of access to data sources}
\usage{
get_groups(
  cookie,
  from = "1900-01-01",
  to = "2999-12-31",
  url = "https://focus.sensingclues.org/"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}
}
\value{
A data frame representing the data sources you have access to.
There are three columns including a data source identifier, a name and the
number of observations made in that group.
}
\description{
Overview of access to data sources
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_groups(cookie)
}
}
