% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{get_heatmap}
\alias{get_heatmap}
\title{Retrieve a heatmap raster}
\usage{
get_heatmap(
  cookie,
  group,
  bounds = NULL,
  from = Sys.Date() - 30,
  to = Sys.Date(),
  aoi = "",
  concepts = NULL,
  resolution = 25,
  type = "observation",
  url = "https://focus.sensingclues.org/"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with \code{\link[=get_groups]{get_groups()}}.}

\item{bounds}{Bounding box coordinates (latitude and longitude) in list(north, east, south, west) format. For example \code{list(north=15, east=10, south=-25, west=50)}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{aoi}{Area of interest.}

\item{concepts}{One or multiple concept definitions, for example \verb{https://sensingclues.poolparty.biz/SCCSSOntology/631}. See \url{https://sensingclues.poolparty.biz/GraphViews/} for all available concepts.}

\item{resolution}{The number of rows and columns of the output raster grid, default is 25.}

\item{type}{The type of data to build the heatmap from, either \code{"observation"} (default) or \code{"track"}.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}
}
\value{
A \code{terra} SpatRaster with \code{resolution} x \code{resolution} grid cells, where each cell holds the number of results collected by the defined group(s) within the given date range and bounds. Returns \code{NULL} when no results are found.
}
\description{
Retrieve a heatmap raster
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
r <- get_heatmap(cookie, group = 'focus-project-1234') # demo group
}
}
