% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Querying Sensing Clues' ontologies}
\usage{
get_hierarchy(url = "https://focus.sensingclues.org/", lang = "en")
}
\arguments{
\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}
}
\value{
Nested list representing multiple ontologies.
}
\description{
Querying Sensing Clues' ontologies
}
\examples{
lst <- get_hierarchy() # you have to be logged in to https://focus.sensingclues.org/ to get access

# inspect a single concept
cncpt <- lst$concepts[[1]]
}
