% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_observations}
\alias{get_observations}
\title{Retrieve observation data}
\usage{
get_observations(
  cookie,
  group,
  bounds = list(north = 90, east = 180, south = -89, west = -179),
  from = Sys.Date() - 30,
  to = Sys.Date(),
  aoi = "",
  filteredConcepts = NULL,
  updateProgress = NULL,
  allAttributes = FALSE,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by login_cluey().}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with get_groups().}

\item{bounds}{Bounding box coordinates (latitude and longitude) in list(north, east, south, west) format}

\item{from}{Start date.}

\item{to}{End date.}

\item{filteredConcepts}{One or multiple concept definitions, for example https://sensingclues.poolparty.biz/SCCSSOntology/186. See https://sensingclues.poolparty.biz/GraphViews/ all available concepts.}

\item{updateProgress}{A function to update a progress bar object, default is NULL.}

\item{allAttributes}{TRUE or FALSE, allows you to collect more attributes for the observations.}

\item{url}{A Sensing Clues URL, default is https://focus.sensingclues.org/.}

\item{lang}{Language in which the concepts are shown, default is english.}
}
\value{
A data frame with all observations collected by the defined group(s), within the given date range.
}
\description{
Retrieve observation data
}
\examples{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_observations(cookie, group = 'focus-project-1234') # demo group
}
