% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_coordinates.R
\name{get_track_coordinates}
\alias{get_track_coordinates}
\title{Retrieve track coordinate data}
\usage{
get_track_coordinates(
  cookie,
  group,
  bounds = NULL,
  from = Sys.Date() - 30,
  to = Sys.Date(),
  aoi = "",
  concepts = NULL,
  updateProgress = NULL,
  total_tracks = 500,
  downsample = TRUE,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with \code{\link[=get_groups]{get_groups()}}.}

\item{bounds}{Bounding box coordinates (latitude and longitude) in list(north, east, south, west) format. For example \code{list(north=15, east=10, south=-25, west=50)}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{aoi}{Area of interest.}

\item{concepts}{One or multiple concept definitions, for example \verb{https://sensingclues.poolparty.biz/SCCSSOntology/631}. See \url{https://sensingclues.poolparty.biz/GraphViews/} for all available concepts.}

\item{updateProgress}{A function to update a progress bar object, default is NULL.}

\item{total_tracks}{The number of tracks to retrieve per page of the query.}

\item{downsample}{A boolean. Allows you to downsample the traces per track.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}
}
\value{
A data frame where each row represents a node in a track. All nodes of all tracks collected by the defined group(s), within the given date range, are returned in the same data frame.
}
\description{
Retrieve track coordinate data
}
\examples{
\dontrun{
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_track_coordinates(cookie, group = 'focus-project-1234') # demo group
}
}
