% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_tracks}
\alias{get_tracks}
\alias{get_track}
\alias{get_track_as_geojson}
\title{Retrieve tracks metadata}
\usage{
get_tracks(
  cookie,
  group,
  bounds = NULL,
  from = Sys.Date() - 30,
  to = Sys.Date(),
  aoi = "",
  patrolType = NULL,
  updateProgress = NULL,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)

get_track(
  cookie,
  trackId,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)

get_track_as_geojson(
  cookie,
  trackId,
  url = "https://focus.sensingclues.org/",
  lang = "en"
)
}
\arguments{
\item{cookie}{A cookie obtained by \code{\link[=login_cluey]{login_cluey()}}.}

\item{group}{One or multiple group identification character string(s), see which groups you have access to with \code{\link[=get_groups]{get_groups()}}.}

\item{bounds}{Bounding box coordinates (latitude and longitude) in list(north, east, south, west) format. For example \code{list(north=15, east=10, south=-25, west=50)}.}

\item{from}{Start date.}

\item{to}{End date.}

\item{aoi}{Area of interest.}

\item{patrolType}{One or multiple concept definitions, for example \verb{https://sensingclues.poolparty.biz/SCCSSOntology/631}. See \url{https://sensingclues.poolparty.biz/GraphViews/} for all available concepts.}

\item{updateProgress}{A function to update a progress bar object, default is NULL.}

\item{url}{A Sensing Clues URL, default is \url{https://focus.sensingclues.org/}.}

\item{lang}{Language in which the concepts are shown, default is English.}

\item{trackId}{Identification character string of a single track.}
}
\value{
A data frame with all tracks collected by the defined group(s), within the given date range.
}
\description{
Retrieve tracks metadata
}
\examples{
\dontshow{if (!is.null(login_cluey())) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cookie <- login_cluey("YOUR_USERNAME", "YOUR_PASSWORD")
df <- get_tracks(cookie, group = 'focus-project-1234') # demo group

# get extra details on a single track
id <- df$entityId[1]
lst <- get_track(cookie, trackId = id)

# get a single track in JSON format
json <- get_track_as_geojson(cookie, trackId = id)
\dontshow{\}) # examplesIf}
}
