% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensingcluesr-package.R
\docType{package}
\name{sensingcluesr-package}
\alias{sensingcluesr}
\alias{sensingcluesr-package}
\title{sensingcluesr: Retrieve Data From Sensing Clues Platform}
\description{
Functions to retrieve data that you have access to on the Sensing Clues platform. This allows you to inspect your data and perform further analyses. The different data types that are available are observations, tracks, map layers and ontologies.
}
\author{
\strong{Maintainer}: Melanie Arp \email{melanie.arp@sensingclues.org}

Authors:
\itemize{
  \item Hugo Koopmans
  \item Chiel Bakkeren
}

}
\keyword{internal}
